code-smells-project/requirements.txt
flask==3.1.1
flask-cors==5.0.1
python-dotenv>=1.0.0
